Made plugin a shared library
