+++ src/compute/cuda_device_memory_allocator.cpp
@@ -26,12 +26,13 @@
  * 
  */
 
-#include "cuda_device_memory_allocator.hpp"
+#include <xmipp4/cuda/compute/cuda_device_memory_allocator.hpp>
 
-#include "cuda_device.hpp"
-#include "cuda_device_queue.hpp"
-#include "cuda_event.hpp"
 #include "default_cuda_device_buffer.hpp"
+
+#include <xmipp4/cuda/compute/cuda_device.hpp>
+#include <xmipp4/cuda/compute/cuda_device_queue.hpp>
+#include <xmipp4/cuda/compute/cuda_event.hpp>
 
 #include <stdexcept>
