+++ src/compute/cuda_device_to_host_transfer.cpp
@@ -26,12 +26,12 @@
  * 
  */
 
-#include "cuda_device_to_host_transfer.hpp"
+#include <xmipp4/cuda/compute/cuda_device_to_host_transfer.hpp>
 
-#include "cuda_error.hpp"
-#include "cuda_device_queue.hpp"
-#include "cuda_device_buffer.hpp"
-#include "cuda_host_memory_allocator.hpp"
+#include <xmipp4/cuda/compute/cuda_error.hpp>
+#include <xmipp4/cuda/compute/cuda_device_queue.hpp>
+#include <xmipp4/cuda/compute/cuda_device_buffer.hpp>
+#include <xmipp4/cuda/compute/cuda_host_memory_allocator.hpp>
 
 #include <xmipp4/core/memory/align.hpp>
 #include <xmipp4/core/compute/host_buffer.hpp>
