+++ src/compute/default_cuda_device_buffer.hpp
@@ -28,7 +28,7 @@
  * 
  */
 
-#include "cuda_device_buffer.hpp"
+#include <xmipp4/cuda/compute/cuda_device_buffer.hpp>
 
 #include <vector>
